// front-coin-wallet/src/app/services/com.b2beyond.api.webwallet-service-user/model/user.ts
/**
 * Api Documentation
 * Api Documentation
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { UserDetails } from './userDetails';


export interface User {
    activationToken?: string;
    active?: boolean;
    authenticationRefreshToken?: string;
    authenticationToken?: string;
    code2FA?: number;
    enabled2FA?: boolean;
    ip?: string;
    password?: string;
    passwordResetToken?: string;
    secretKey?: string;
    userDetails?: UserDetails;
    userId?: string;
    valid?: boolean;
    websiteId?: number;
}